+++ NEWS.md
@@ -1,3 +1,5 @@
+# roclang 0.2.0
+
 # roclang 0.1.4
 
 ## Enhancements: